refactor: Separate PuyoType and PuyoAttribute from puyo.ts (#19)
