+++ src/reducers/puyoAppSlice.ts
@@ -9,16 +9,16 @@ import { type BoardEditMode, HowToEditBoard } from '../logics/BoardEditMode';
 import { getBoostArea } from '../logics/BoostArea';
 import type { OptimizationTarget } from '../logics/OptimizationTarget';
 import { PuyoCoord } from '../logics/PuyoCoord';
-import { Simulator } from '../logics/Simulator';
-import type { TraceMode } from '../logics/TraceMode';
-import { screenshotBoardId } from '../logics/boards';
-import type { ChainDamage } from '../logics/damage';
 import {
   PuyoType,
   toChanceColoredType,
   toNormalColoredType,
   toPlusColoredType
-} from '../logics/puyo';
+} from '../logics/PuyoType';
+import { Simulator } from '../logics/Simulator';
+import type { TraceMode } from '../logics/TraceMode';
+import { screenshotBoardId } from '../logics/boards';
+import type { ChainDamage } from '../logics/damage';
 import { type ExplorationResult, SolutionMethod } from '../logics/solution';
 import { INITIAL_PUYO_APP_STATE, type PuyoAppState } from './PuyoAppState';
 import { reflectBoardInSimulator } from './internal/reflectBoardInSimulator';
